Financial-Analyst/pyproject.toml
[project]
name = "financial-analyst"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "firecrawl-py==2.16.3",
    "google-adk>=1.18.0",
    "google-genai>=1.31.0",
    "litellm>=1.75.8",
    "python-dotenv>=1.1.1",
    "yfinance>=0.2.65",
    "fastapi==0.118.3",
    "pydantic==2.10.6",
]

[dependency-groups]
dev = [
    "watchdog>=6.0.0",
]

[tool.uv]
override-dependencies=[
    "pydantic==2.10.6"
]
